+++ .dockerignore
@@ -1,0 +1,4 @@
+# comment
+drafts/
+__pycache__/
+venv/
