+++ README.md
@@ -34,12 +34,17 @@ MQTT-USERNAME = "mqtt-test"
 MQTT-PASSWORD = "mqtt-test"
 ```
 
+## Docker setup
+```
+docker pull rabbitmq
+```
+
 ## App1
 Generates a random number between [1 and 100] every [1 to 30] seconds and publishes to broker with topic named `app1/randint`.
 
 Run app:
 ```
-python app1.py
+python src/app1.py
 ```
 ## App2
 Subscribes to topic `app1/randint` and compute average every [1, 5, 30] minutes,
@@ -47,7 +52,7 @@ and publishes accordingly to broker with topics named `app2/stat_1m`, `app2/stat
 
 Run app:
 ```
-python app2.py
+python src/app2.py
 ```
 
 ## App3
@@ -55,7 +60,7 @@ Subscribes to topics `app2/stat_1m`, `app2/stat_5m`, `app2/stat_30m`, and prints
 
 Run app:
 ```
-python app3.py
+python src/app3.py
 ```
 
 # Demo
