+++ docker/Dockerfile.app1
@@ -1,0 +1,13 @@
+FROM python:3.10-slim
+
+WORKDIR /usr/src/app
+
+COPY requirements.txt ./
+RUN pip install --upgrade pip
+RUN pip install --no-cache-dir -r requirements.txt
+
+COPY . .
+
+EXPOSE 1883 11883
+
+CMD [ "python", "app1.py" ]
