+++ docker/Dockerfile.broker
@@ -1,0 +1,12 @@
+FROM rabbitmq:3.11.2-alpine
+
+RUN rabbitmq-plugins enable --offline rabbitmq_mqtt
+
+# ADD init.sh /init.sh
+# RUN chmod +x /init.sh
+
+EXPOSE 1883 11883
+
+
+CMD ["rabbitmq-server"]
+# RUN /init.sh
